refactor: extract route handlers into Fastify plugins (#8)
Split monolithic index.ts files in both services into modular Fastify
plugin route files, improving navigability and maintainability.

listing-processor (1025 → 70 lines in index.ts):
- 5 helper modules: flash, render, odoo, listing, cache
- 6 route plugins: dashboard, products, queue, settings, ai, categories

upload-api (537 → 51 lines in index.ts):
- 3 helper modules: idempotency, ebay-errors, upload-logic
- 7 route plugins: health, connection, category, upload, verify, revise, status
